Group_7/requirements.txt
streamlit
duckdb
pandas
numpy
plotly
requests
faiss-cpu
sentence-transformers
pyarrow
python-dotenv
llama-index
chromadb
transformers
torch
accelerate
huggingface-hub
openai
tqdm
scikit-learn
matplotlib
seaborn
jupyter
notebook
